xmhf: libxmhfgeec: fix XMHFGEEC_SLAB_UAPICAP_MASK definition to use u32 instead of u8; successful test.

+++ xmhf-libs/libxmhfgeec/include/xmhfgeec.h
@@ -149,7 +149,7 @@ typedef struct {
 
 
 #define XMHFGEEC_SLAB_CALLCAP_MASK(x)               (1UL << x)
-#define XMHFGEEC_SLAB_UAPICAP_MASK(x)               (u8)(1 << x)
+#define XMHFGEEC_SLAB_UAPICAP_MASK(x)               (1UL << x)
 #define XMHFGEEC_SLAB_MEMGRANTREADCAP_MASK(x)       (1UL << x)
 #define XMHFGEEC_SLAB_MEMGRANTWRITECAP_MASK(x)      (1UL << x)
